server/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Redis and Celery
redis==5.2.0
celery==5.4.0
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Storage (S3/MinIO)
boto3==1.35.20
minio==7.2.7

# Audio processing
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4

# HTTP clients
httpx==0.27.2
aiohttp==3.10.11

# Model providers
fal-client==0.4.0
replicate==0.34.0

# Security
clamd==1.0.2
python-magic==0.4.27

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.40b0
sentry-sdk[fastapi]>=2.0.0
prometheus-client>=0.20.0

# Payments
stripe==5.5.0

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.23.7

# Utilities
python-dotenv==1.0.1
pytz==2024.2

